added support in ObjectFieldTemplate to support side-by-side layout w/ burnettk
